Push down limit statements on cottontail (ref #357)
