+++ cottontail-adapter/src/main/java/org/polypheny/db/adapter/cottontail/rules/CottontailRules.java
@@ -40,7 +40,8 @@ public class CottontailRules {
                 new CottontailTableModificationRule( out, relBuilderFactory ),
                 new CottontailProjectRule( out, relBuilderFactory ),
                 new CottontailFilterRule( out, relBuilderFactory ),
-                new CottontailSortAndProjectRule( out, relBuilderFactory )
+                new CottontailSortAndProjectRule( out, relBuilderFactory ),
+                new CottontailLimitRule( out, relBuilderFactory )
         );
     }
